누구야1 게시글 등록 기능 연동

+++ src/main/java/io/weeks/nuguya/dto/CrawlingDto.java
@@ -9,6 +9,6 @@ public class CrawlingDto {
 
     private String requestUrl;
 
-    private int crawlingNum;
+    private Long crawlingNum;
 
 }
